+++ src/test/java/chess/controller/ChessSpringControllerTest.java
@@ -105,51 +105,4 @@ public class ChessSpringControllerTest {
                 .queryParam("id", "1"))
                 .andExpect(content().json(responseDto.toJson()));
     }
-
-//    @DisplayName("chess Test- GET")
-//    @Test
-//    void get_chess() throws Exception {
-//        final ChessBoardDto chessBoardDto = ChessBoardDto.from(new Board().getPiecesByPosition());
-//        given(chessGameService.getBoard()).willReturn(ChessBoardDto.from(new Board().getPiecesByPosition()));
-//
-//        mockMvc.perform(get("/chess"))
-//                .andExpect(view().name("index"))
-//                .andExpect(model().attribute("boardDto", chessBoardDto));
-//    }
-
-//    @DisplayName("move POST 요청 테스트")
-//    @Test
-//    void post_move() throws Exception {
-//        final String requestString = "a2 a4";
-//        final ResponseDto responseDto = new ResponseDto(302, "");
-//        given(chessGameService.move("a2", "a4")).willReturn(responseDto);
-//        mockMvc.perform(post("/move")
-//                .content(requestString))
-//                .andExpect(content().json(responseDto.toJson()));
-//
-//    }
-
-//    @DisplayName("end GET 요청 테스트")
-//    @Test
-//    void get_end() throws Exception {
-//        ResponseDto responseDto = new ResponseDto(200, "");
-//        given(chessGameService.end()).willReturn(new ResponseDto(200, ""));
-//
-//        mockMvc.perform(get("/end"))
-//                .andExpect(content().json(responseDto.toJson()));
-//    }
-
-//    @DisplayName("result GET 요청 테스트")
-//    @Test
-//    void get_result() throws Exception {
-//        final ResultDto resultDto = ResultDto.of(36, 37, Winner.BLACK);
-//
-//        given(chessGameService.statusOfWhite()).willReturn(36.0);
-//        given(chessGameService.statusOfBlack()).willReturn(37.0);
-//        given(chessGameService.findWinner()).willReturn(Winner.BLACK);
-//
-//        mockMvc.perform(get("/chess-result"))
-//                .andExpect(view().name("result"))
-//                .andExpect(model().attribute("result", resultDto));
-//    }
 }
