+++ src/test/java/chess/dao/MockBoardDao.java
@@ -1,10 +1,12 @@
 package chess.dao;
 
-import chess.domain.ChessGame2;
+import chess.domain.ChessGame;
 import chess.domain.Color;
 import chess.domain.Room;
 import chess.domain.board.Board;
+import chess.dto.RoomDto;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 public class MockBoardDao implements BoardDao {
@@ -33,9 +35,9 @@ public class MockBoardDao implements BoardDao {
     }
 
     @Override
-    public Long save(ChessGame2 chessGame2) {
-        final Board board = chessGame2.getBoard();
-        final Room room = chessGame2.getRoom();
+    public Long save(ChessGame chessGame) {
+        final Board board = chessGame.getBoard();
+        final Room room = chessGame.getRoom();
         fakeBoard.put(1L, new FakeBoard(board.getTurn().name(), room.getTitle(), room.getPassword()));
         return 1L;
     }
@@ -47,5 +49,15 @@ public class MockBoardDao implements BoardDao {
         final String password = fakeBoard.getPassword();
         this.fakeBoard.remove(boardId);
         this.fakeBoard.put(boardId, new FakeBoard(turn.name(), title, password));
+    }
+
+    @Override
+    public List<RoomDto> findAllRooms() {
+        return null;
+    }
+
+    @Override
+    public String findPasswordById(Long boardId) {
+        return null;
     }
 }
